Refactoring: object_hook_trusted.py and json_dumps.py
