+++ blaze/utils.py
@@ -1,8 +1,7 @@
 from __future__ import absolute_import, division, print_function
-import types
 
 import datetime
-from collections import Iterator, Callable
+from collections import Iterator
 from itertools import islice, product
 import os
 import re
@@ -12,7 +11,6 @@ try:
 except ImportError:
     from toolz import nth, unique, concat
 
-from datashape import Mono, DataShape
 import numpy as np
 # these are used throughout blaze, don't remove them
 from odo.utils import tmpfile, filetext, filetexts, raises, keywords, ignoring
@@ -22,9 +20,9 @@ import sqlalchemy as sa
 
 # Imports that replace older utils.
 from .compatibility import map, zip
-from .dispatch import dispatch
 
-from .object_hook import object_hook
+from .object_hook_trusted import object_hook_trusted as object_hook
+from .json_dumps import json_dumps
 
 
 def nth_list(n, seq):
@@ -154,54 +152,6 @@ def listpack(x):
         return x
     else:
         return [x]
-
-
-@dispatch(datetime.datetime)
-def json_dumps(dt):
-    if dt is pd.NaT:
-        # NaT has an isoformat but it is totally invalid.
-        # This keeps the parsing on the client side simple.
-        s = 'NaT'
-    else:
-        s = dt.isoformat()
-        if not dt.tzname():
-            s += 'Z'
-
-    return {'__!datetime': s}
-
-
-@dispatch(frozenset)
-def json_dumps(ds):
-    return {'__!frozenset': list(ds)}
-
-
-@dispatch(datetime.timedelta)
-def json_dumps(ds):
-    return {'__!timedelta': ds.total_seconds()}
-
-
-@dispatch(Mono)
-def json_dumps(m):
-    return {'__!mono': str(m)}
-
-
-@dispatch(DataShape)
-def json_dumps(ds):
-    return {'__!datashape': str(ds)}
-
-
-@dispatch(types.BuiltinFunctionType)
-def json_dumps(f):
-    return {'__!builtin_function': f.__name__}
-
-
-@dispatch(Callable)
-def json_dumps(f):
-    # let the server serialize any callable - this is only used for testing
-    # at present - do the error handling when json comes from client so in
-    # object_hook, catch anything that is not pandas_numpy
-    fcn = ".".join([f.__module__, f.__name__])
-    return {'__!numpy_pandas_function': fcn}
 
 
 def normalize(s):
